+ Created basic interface for a battle

+++ src/main/model/graphics/DefaultScene.java
@@ -1,8 +1,8 @@
 package main.model.graphics;
 
 public abstract class DefaultScene {
-    public final static int FINAL_WIDTH = 1040;
-    public final static int FINAL_HEIGHT = 820;
+    public final static double FINAL_WIDTH = 1040;
+    public final static double FINAL_HEIGHT = 820;
 
     protected abstract void initializeGraphics();
 }
